requirements.txt
a2wsgi==1.10.10
alembic==1.17.1
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
flasgger==0.9.7.1
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
h11==0.16.0
httptools==0.7.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.1.0
marshmallow-sqlalchemy==1.4.2
mistune==3.1.4
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
